requirements.txt
ktrain
pdfplumber
pdfminer.six
tika
textwrap3
python-docx